requirements.txt
boto3>=1.26.0
confluent-kafka>=2.1.1
fastavro>=1.7.0
python-dotenv>=1.0.0
PyMuPDF>=1.22.5
pymongo>=4.4.1
langchain>=0.0.200
llama-parse>=0.0.27
typing-extensions>=4.5.0
